# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Certif",
        "description": "Api de l'appli PENC",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/displaymessage": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Retourne la liste des sujets avec leurs messages associés",
                "description": "Retourne la liste des sujets avec leurs messages associés.",
                "operationId": "6d879defb5f48e74a1e221a53bf3ced9",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An unexpected error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sendmessage": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Crée un nouveau message dans un sujet spécifié",
                "description": "Crée un nouveau message dans un sujet spécifié.",
                "operationId": "514defc4d6d1297e61e693997583ab54",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_content",
                                    "topic_id"
                                ],
                                "properties": {
                                    "message_content": {
                                        "type": "string",
                                        "example": "Contenu du message"
                                    },
                                    "topic_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Votre message a bien été créé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête incorrecte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Un administrateur ne peut pas envoyer de message."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Une erreur est survenue lors de la création du message."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/updatespecificmessage/{message}": {
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Met à jour un message existant",
                "description": "Met à jour un message existant.",
                "operationId": "1d8f083ca53736d8103878f424c9dfbf",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "ID du message à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_content",
                                    "user_id",
                                    "topic_id"
                                ],
                                "properties": {
                                    "message_content": {
                                        "type": "string",
                                        "example": "Nouveau contenu du message"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "topic_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Modifications effectuées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Modifications effectuées avec succès."
                                        },
                                        "resultat": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ce message ne vous appartient pas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne. Veuillez réessayer."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Message d'erreur spécifique"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/deletespecificmessage/{message}": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Supprime un message existant",
                "description": "Supprime un message existant.",
                "operationId": "c237358655bcd0fe27f0d0146da473b9",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "ID du message à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message supprimé avec succès par l'administrateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Le message a été supprimé avec succès par l'administrateur."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vous n'êtes pas autorisé à supprimer ce message."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur est survenue lors de l'authentification."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}